requirements.txt
crewai
crewai_tools
python-dotenv
fastapi
uvicorn
websockets
litellm
openai
